Type the party item pool count unsigned
Every retail comparison against the count is unsigned (jb/jae/jbe), and
the source cast it back to unsigned at each loop.

+++ src/wiz8/local_code/ConditionsAndEnchantments.cpp
@@ -282,7 +282,7 @@ void SanitizeLoadedItems(void)
             }
         }
     }
-    for (index = 0; index < static_cast<unsigned int>(g_status.party_item_count_1791); ++index) {
+    for (index = 0; index < g_status.party_item_count_1791; ++index) {
         if (g_status.party_item_pool_0021[index].iItemNo != -1) {
             NormalizeItemQuantityKind(&g_status.party_item_pool_0021[index]);
         }
